requirements.txt
absl-py==0.9.0
amqp==2.6.0
astor==0.8.1
astunparse==1.6.3
billiard==3.6.3.0
cachetools==4.1.0
celery==4.4.5
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dnspython==1.16.0
easydict==1.9
Flask==1.1.2
Flask-Script==2.0.6
future==0.18.2
gast==0.2.2
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kombu==4.6.10
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
monotonic==1.5
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.3
Pillow==6.2.0
pkg-resources==0.0.0
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.2
requests==2.23.0
requests-oauthlib==1.3.0
rq==1.4.2
rsa==4.0
scipy==1.1.0
seaborn==0.9.0
six==1.15.0
tensorboard==2.0.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-gpu==2.2.0
termcolor==1.1.0
urllib3==1.25.9
vine==1.3.0
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
zipp==3.1.0
